Renames broadcast and removes deprecated files
