+++ Demo/DemoSnapshotTests.swift
@@ -39,10 +39,6 @@ class ComponentViewTests: FBSnapshotTestCase {
         snapshot(.broadcast)
     }
 
-    func testBroadcastContainer() {
-        snapshot(.broadcastContainer)
-    }
-
     func testButton() {
         snapshot(.button)
     }
